+++ dlls/combat.cpp
@@ -1951,7 +1951,7 @@ Go to the trouble of combining multiple pellets into a single damage call.
 This version is used by Monsters.
 ================
 */
-void CBaseEntity::FireBullets( ULONG cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker )
+void CBaseEntity::FireBullets( unsigned int cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker )
 {
 	static int tracerCount;
 	TraceResult tr;
@@ -1967,7 +1967,7 @@ void CBaseEntity::FireBullets( ULONG cShots, Vector vecSrc, Vector vecDirShootin
 
 	UTIL_MuzzleLight(vecSrc);
 
-	for( ULONG iShot = 1; iShot <= cShots; iShot++ )
+	for( unsigned int iShot = 1; iShot <= cShots; iShot++ )
 	{
 		// get circular gaussian spread
 		float x, y, z;
@@ -2025,7 +2025,7 @@ Go to the trouble of combining multiple pellets into a single damage call.
 This version is used by Players, uses the random seed generator to sync client and server side shots.
 ================
 */
-Vector CBaseEntity::FireBulletsPlayer( ULONG cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
+Vector CBaseEntity::FireBulletsPlayer( unsigned int cShots, Vector vecSrc, Vector vecDirShooting, Vector vecSpread, float flDistance, int iBulletType, int iTracerFreq, int iDamage, entvars_t *pevAttacker, int shared_rand )
 {
 	TraceResult tr;
 	Vector vecRight = gpGlobals->v_right;
@@ -2041,7 +2041,7 @@ Vector CBaseEntity::FireBulletsPlayer( ULONG cShots, Vector vecSrc, Vector vecDi
 	DamageInfo damageInfo{(float)iDamage, DMG_BULLET};
 	damageInfo.SetGibPolicy(GIB_NEVER);
 
-	for( ULONG iShot = 1; iShot <= cShots; iShot++ )
+	for( unsigned int iShot = 1; iShot <= cShots; iShot++ )
 	{
 		//Use player's random seed.
 		// get circular gaussian spread
