+++ dlls/globals.cpp
@@ -27,9 +27,9 @@
 #include "util.h"
 #include "cbase.h"
 
-DLL_GLOBAL ULONG		g_ulFrameCount;
-DLL_GLOBAL ULONG		g_ulModelIndexEyes;
-DLL_GLOBAL ULONG		g_ulModelIndexPlayer;
+DLL_GLOBAL unsigned int		g_ulFrameCount;
+DLL_GLOBAL unsigned int		g_ulModelIndexEyes;
+DLL_GLOBAL unsigned int		g_ulModelIndexPlayer;
 DLL_GLOBAL Vector		g_vecAttackDir;
 DLL_GLOBAL int			g_iSkillLevel;
 DLL_GLOBAL bool			gDisplayTitle;
