+++ dlls/player.cpp
@@ -52,7 +52,7 @@
 
 // #define DUCKFIX
 
-extern DLL_GLOBAL ULONG g_ulModelIndexPlayer;
+extern DLL_GLOBAL unsigned int g_ulModelIndexPlayer;
 extern DLL_GLOBAL bool g_fGameOver;
 bool gEvilImpulse101;
 extern DLL_GLOBAL int g_iSkillLevel;
